ai-platform/services/api_gateway/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
PyJWT==2.8.0
redis==5.0.6
confluent-kafka==2.4.0
python-dotenv==1.0.1
